b_fix : FIXME : Texture or Not? #70
