+++ etc/rt/texture.rt
@@ -5,6 +5,7 @@ pl 300,300,300 -.1,-.1,-.1 165,150,180
 sp 60,60,60 120 255,255,255
 sp 200,10,10 120 255,0,0
 sp 10,250,10 60 0,0,255
+cy 0,0,0     0,1,1 130       200       10,0,255
 sp 10,10,160 30 0,255,0
 sp -100,200,200 200 255,255,0
 tx moon.ppm 1,3,5
