+++ include/bonus/minirt_bonus.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minirt_bonus.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yongjule <yongjule@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/15 14:49:32 by ghan              #+#    #+#             */
-/*   Updated: 2021/12/16 12:03:51 by yongjule         ###   ########.fr       */
+/*   Updated: 2021/12/16 16:09:27 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -149,6 +149,7 @@ int			circle_cn_shadow(double *ray, t_pt_info *pt_i,
 				t_cn *cn, double r_size);
 int			cy_shadow(double *ray, t_pt_info *pt_i, t_cy *cy, double r_size);
 int			cn_shadow(double *ray, t_pt_info *pt_info, t_cn *cn, double r_size);
+int			get_shadow(t_l_lst *cur_lp, t_obj_lst *hd, t_pt_info *pt_info);
 int			intersect_pl(double *ray, t_pt_info *pt_info, t_pl *pl);
 int			intersect_sph(double *ray, t_pt_info *pt_info, t_sph *sph);
 int			intersect_cy(double *ray, t_pt_info *pt_info, t_cy *cy);
@@ -157,9 +158,9 @@ int			intersect_cn(double *ray, t_pt_info *pt_info, t_cn *cn);
 int			intersect_cn_circle(double *ray, t_pt_info *pt_i, t_cn *cn);
 double		meet_pl(double *ray, double *o_vect);
 double		meet_sph(double *ray, double *origin, t_sph *sph, double r_size);
-int			get_shadow(t_l_lst *cur_lp, t_obj_lst *hd, t_pt_info *pt_info);
 int			pl_shadow(double *ray, t_pt_info *pt_info,
 				t_pl *pl, double r_size);
+int			shoot_ray(t_rt *rt, double vs_x, double vs_y);
 
 /* Hook */
 void		hook_minirt(t_rt *rt);
