+++ src/bonus/render/apply_texture_bonus.c
@@ -3,26 +3,42 @@
 /*                                                        :::      ::::::::   */
 /*   apply_texture_bonus.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ghan <ghan@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/15 14:39:27 by ghan              #+#    #+#             */
-/*   Updated: 2021/12/15 16:29:10 by ghan             ###   ########.fr       */
+/*   Updated: 2021/12/16 15:42:53 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt_bonus.h"
 
-void    apply_checker(t_pt_info *pt_info)
+void	apply_checker(t_pt_info *pt_info)
 {
-    double  sines;
+	double  sines;
+	double  div;
 
-    sines = sin(pt_info->pt[X] / 100) * sin(pt_info->pt[Y] / 100)
-        * sin(pt_info->pt[Z] / 100);
-    if (sines > 0)
-        ft_bzero(pt_info->color, sizeof(int) * 3);
+	div = 100;
+	if (pt_info->type == PLANE)
+		div = 400;
+	else if (pt_info->type == SPHERE)
+		div = pt_info->obj.sph->radius / 4;
+	else if (pt_info->type == CYLINDER)
+		div = pt_info->obj.cy->radius / 4;
+	else if (pt_info->type == CONE)
+		div = pt_info->obj.cn->radius / 4;
+	sines = sin(pt_info->pt[X] / div) * sin(pt_info->pt[Y] / div)
+		* sin(pt_info->pt[Z] / div);
+	if (!signbit(sines))
+		ft_bzero(pt_info->color, sizeof(int) * 3);
+	else
+	{
+		pt_info->color[R] = 255;
+		pt_info->color[G] = 255;
+		pt_info->color[B] = 255;
+	}
 }
 
-void    apply_texture(t_pt_info *pt_info)
+void	apply_texture(t_pt_info *pt_info)
 {
-    pt_info->is_txt = CUSTOM_TXT;
+	pt_info->is_txt = CUSTOM_TXT;
 }
